475: fix funnel tests to work with worker
