+++ src/global/class/Point.js
@@ -8,14 +8,13 @@ export class Point {
   constructor(x, y) {
     this.x = x;
     this.y = y;
-    this.clearancePoint = null;
   }
 
 
   fromObject(o) {
     this.x = o.x;
     this.y = o.y;
-    this.clearancePoint = o.clearancePoint ? (new Point().fromObject(o.clearancePoint)) : null;
+    if (o.clearancePoint) this.clearancePoint = new Point().fromObject(o.clearancePoint);
     return this;
   }
